pyproject.toml
[project]
name = "version-control-helper-mcp"
version = "0.1.0"
description = "MCP server for code version control operations"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.26.0",
    "gitpython>=3.1.46",
    "pygithub>=2.8.1",
    "pydantic>=2.0.0",
]

[project.scripts]
version-control-helper-mcp = "version_control_helper_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/version_control_helper_mcp"]

[dependency-groups]
dev = []
